move formatting logic to external fn
