+++ packages/backend/src/core/liveness/types/LivenessConfig.ts
@@ -1,0 +1,16 @@
+import { FunctionCallQueryParams, TransferQueryParams } from '@l2beat/shared'
+import { LivenessType, ProjectId } from '@l2beat/shared-pure'
+
+export interface LivenessFunctionCall extends FunctionCallQueryParams {
+  type: LivenessType
+}
+
+export interface LivenessTransfer extends TransferQueryParams {
+  type: LivenessType
+}
+
+export interface LivenessConfig {
+  projectId: ProjectId
+  transfers?: LivenessTransfer[]
+  functionCalls?: LivenessFunctionCall[]
+}
