+++ packages/backend/src/core/liveness/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './formatFunctionCallsQueryResult'
+export * from './formatTransfersQueryResult'
